services/verification/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
dnspython==2.7.0
pydantic==2.10.3
